requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
opencv-python
Pillow
tensorflow
keras
imblearn
joblib